refactor: revert extraction of VariableFetcher and migrate JudgeAssertj into VariableAssertj
